munin_node_os.t: remove tests related to possible_to_signal_process
This test assumes that if we are root we can kill pid1 with signal 0, and if
we are not root, we cannot.

This is a good assumption where pid1 is running as root, but in docker it is
run by the user itself, so our build fails as it can kill the pid1 with
signal 0.

The underlying code is so simple that instead of fixing this tests (it would
be better to stat() the /proc/1 directory, then compare the UID with the
current user's UID), the tests were removed instead.